refactor: item menu takes deletePrompt and editPath props, no longer builds them internally

+++ src/components/Shared/ItemMenu.tsx
@@ -3,7 +3,6 @@
 import { Delete, Edit, MoreHoriz } from "@mui/icons-material";
 import { IconButton, Menu, MenuItem, SxProps } from "@mui/material";
 import { ReactNode } from "react";
-import { NavigationPaths } from "../../constants/common.constants";
 import { useTranslate } from "../../hooks/common.hooks";
 import { isRenderable } from "../../utils/common.utils";
 import GhostButton from "./GhostButton";
@@ -16,9 +15,9 @@ interface Props {
   canEdit?: boolean;
   children?: ReactNode;
   deleteItem?: (id: number) => void;
+  deletePrompt?: string;
+  editPath?: string;
   itemId: number;
-  itemType: string;
-  name?: string;
   prependChildren?: boolean;
   setAnchorEl: (el: null | HTMLElement) => void;
   variant?: "ghost" | "default";
@@ -31,9 +30,9 @@ const ItemMenu = ({
   canEdit,
   children,
   deleteItem,
+  deletePrompt,
+  editPath,
   itemId,
-  itemType,
-  name,
   prependChildren,
   setAnchorEl,
   variant,
@@ -49,9 +48,6 @@ const ItemMenu = ({
 
   const handleClose = () => setAnchorEl(null);
 
-  const editItemPath = `/${itemType}/${name ? name : itemId}${
-    NavigationPaths.Edit
-  }`;
   const editIconStyles = {
     marginRight: 1,
     transform: "rotateY(180deg) translateY(2px)",
@@ -60,6 +56,9 @@ const ItemMenu = ({
     transform: `translateY(${variant === "ghost" ? 4 : 1}px)`,
   };
   const Button = variant === "ghost" ? GhostButton : IconButton;
+
+  const showEditButton = canEdit && editPath;
+  const showDeleteButton = canDelete && deleteItem && deletePrompt;
 
   return (
     <>
@@ -88,22 +87,18 @@ const ItemMenu = ({
       >
         {prependChildren && children}
 
-        {canEdit && (
+        {showEditButton && (
           <MenuItem>
-            <Link sx={{ color: "inherit" }} href={editItemPath}>
+            <Link sx={{ color: "inherit" }} href={editPath}>
               <Edit fontSize="small" sx={editIconStyles} />
               {t("actions.edit")}
             </Link>
           </MenuItem>
         )}
 
-        {canDelete && deleteItem && (
+        {showDeleteButton && (
           <MenuItem
-            onClick={() =>
-              // FIXME: Item type is currently showing as plural - is also used above in edit path
-              window.confirm(t("prompts.deleteItem", { itemType })) &&
-              deleteItem(itemId)
-            }
+            onClick={() => window.confirm(deletePrompt) && deleteItem(itemId)}
           >
             <Delete fontSize="small" sx={{ marginRight: 1 }} />
             {t("actions.delete")}
